Add configuration options for TCP sender #62
The TCP protocol nature allows to specify read/connection timeouts and keep-alive. The sender requires an established TCP connection that is open and valid from both ends. If a TCP connection times out (broken pipe from the other end) it failed before and dropped the log message. To avoid this, another option was introduced: deliveryAttempts. This option allows to (re)deliver the message until it was sent sucessfully.