Do not send source/location fields if source location is not available #67
log4j, log4j2, java.util.logging and logback behave differently if no source/location details are available. A breaking behavior for graylog users is log4j's "?" if no location is available. logstash-gelf checks for the known unknown-location markers and omits the source fields if the location details are not available.