Rethrow Exceptions in log4j2 GelfLogAppender #78
The GelfLogAppender for log4j2 is now able to throw exceptions when setting ignoreExceptions to false in the appender configuration. Exception propagation is useful for connection errors and exceptions in general which occur during log event processing. Propagating Exceptions can be used together with the FailoverAppender to use secondary appenders. Setting ignoreExceptions to false will cause the log invocation to fail when not using together with FailoverAppender.