Close channel in GelfTCPSender.connect() #107
Always close a channel, if present when attempting connection. Closing the channel is required if the connection failed due to a broken pipe or any other intermediate error which leaves the client channel not closed. Previously, a wrong check to Channel.isOpen() was verifying the wrong state and caused IllegalStateException that are not catched.