Fix a memory leak in GelfUDPSender
ThreadLocal instances should be defined as `static` so that
the ThreadLocal pool is shared between all instances of a class.
With a non-static ThreadLocal instance, each instance of the
GelfUDPSender will spin up a new ThreadLocal instance, which
will not know about the other ThreadLocal instances, and
will reinitialize the buffer even if another instance in the
same thread has already done so.